chore(IdentityHeader): RHICOMPL-3761 move IdentityHeader to lib
