+++ lib/insights.rb
@@ -5,6 +5,7 @@ require 'insights/api/common/satellite_compensation/middleware'
 require 'insights/api/common/logger_with_audit'
 require 'insights/api/common/platform'
 require 'insights/api/common/host_inventory'
+require 'insights/api/common/identity_header'
 
 # Module to hold every platform-related code
 module Insigths
