+++ test/controllers/concerns/authentication_test.rb
@@ -327,7 +327,7 @@ class AuthenticationTest < ActionController::TestCase
         {
           'identity': {
             'org_id': '1234',
-            'auth_type': IdentityHeader::CERT_AUTH
+            'auth_type': Insights::Api::Common::IdentityHeader::CERT_AUTH
           },
           'entitlements':
           {
