+++ test/controllers/v1/rules_controller_test.rb
@@ -16,7 +16,7 @@ module V1
         {
           'identity': {
             'org_id': account.org_id,
-            'auth_type': IdentityHeader::CERT_AUTH
+            'auth_type': Insights::Api::Common::IdentityHeader::CERT_AUTH
           },
           'entitlements':
           {
@@ -41,7 +41,7 @@ module V1
         {
           'identity': {
             'org_id': account.org_id,
-            'auth_type': IdentityHeader::CERT_AUTH
+            'auth_type': Insights::Api::Common::IdentityHeader::CERT_AUTH
           },
           'entitlements':
           {
