+++ test/factories/account.rb
@@ -4,7 +4,7 @@ FactoryBot.define do
   factory :account do
     sequence(:org_id) { |n| format('%05<num>d', num: n) }
     identity_header do
-      IdentityHeader.new(
+      Insights::Api::Common::IdentityHeader.new(
         Base64.encode64({
           identity: {
             org_id: org_id
